test: use `getPageAndLoaderData()` helper function
