+++ packages/@expo/cli/e2e/__tests__/export/static-loader.test.ts
@@ -6,7 +6,7 @@ import {
   RUNTIME_EXPO_START,
   setupServer,
 } from '../../utils/runtime';
-import { findProjectFiles, getHtml } from '../utils';
+import { findProjectFiles, getPageAndLoaderData } from '../utils';
 
 runExportSideEffects();
 
@@ -109,43 +109,44 @@ describe.each(
     expect(data).toEqual({ foo: 'bar' });
   });
 
-  it('loader endpoint returns `null` for `undefined` loader data', async () => {
-    const response = await server.fetchAsync('/_expo/loaders/nullish/undefined');
-    expect(response.status).toBe(200);
-    const data = await response.json();
-    expect(data).toBeNull();
-  });
+  it.each(getPageAndLoaderData('/nullish/undefined'))(
+    'returns `null` for `undefined` loader data for $url ($name)',
+    async ({ getData, url }) => {
+      const response = await server.fetchAsync(url);
+      expect(response.status).toBe(200);
+      const data = await getData(response);
+      expect(data).toBeNull();
+    }
+  );
 
-  it('loader endpoint returns `null` for `null` loader data', async () => {
-    const response = await server.fetchAsync('/_expo/loaders/nullish/null');
-    expect(response.status).toBe(200);
-    const data = await response.json();
-    expect(data).toBeNull();
-  });
+  it.each(getPageAndLoaderData('/nullish/null'))(
+    'returns `null` for `null` loader data for $url ($name)',
+    async ({ getData, name, url }) => {
+      const response = await server.fetchAsync(url);
+      expect(response.status).toBe(200);
+      const data = await getData(response);
 
-  it.each([
-    {
-      name: 'loader endpoint',
-      url: '/_expo/loaders/request',
-      getData: (response: Response) => {
-        return response.json();
-      },
-    },
-    {
-      name: 'page',
-      url: '/request',
-      getData: async (response: Response) => {
-        const html = getHtml(await response.text());
-        return JSON.parse(html.querySelector('[data-testid="loader-result"]')!.textContent);
-      },
-    },
-  ])('$name $url does not receive `Request` object', async ({ getData, url }) => {
-    const response = await server.fetchAsync(url);
-    expect(response.status).toBe(200);
-    const data = await getData(response);
+      // NOTE(@hassankhan): For HTML pages, the fixture component converts `null` to the
+      // string `NULL` for display (see `nullish/[value].tsx`). The loader endpoint
+      // returns the raw `null` value.
+      if (name === 'page') {
+        expect(data).toEqual('NULL');
+      } else {
+        expect(data).toBeNull();
+      }
+    }
+  );
 
-    expect(data.url).toBeNull();
-    expect(data.method).toBeNull();
-    expect(data.headers).toBeNull();
-  });
+  it.each(getPageAndLoaderData('/request'))(
+    'does not receive `Request` object for $url ($name)',
+    async ({ getData, url }) => {
+      const response = await server.fetchAsync(url);
+      expect(response.status).toBe(200);
+      const data = await getData(response);
+
+      expect(data.url).toBeNull();
+      expect(data.method).toBeNull();
+      expect(data.headers).toBeNull();
+    }
+  );
 });
